requirements.txt
pytest~=8.3.5
selenium~=4.33.0
PyYAML~=6.0.3
requests~=2.32.5
allure-python-commons~=2.14.2